pyproject.toml
[project]
name = "ais-nlp10s"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.21",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.3.1",
    "langchain-openai>=0.3.12",
    "pymupdf>=1.25.5",
    "pypdf>=5.4.0",
    "python-docx>=1.1.2",
    "ruff>=0.11.4",
    "semanticscholar>=0.10.0",
    "sentence-transformers>=4.0.2",
    "typing-extensions>=4.13.1",
    "wikipedia>=1.4.0",
]
